raio-x-digital/python_agent/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
apify_client==1.12.2
apify_shared==1.5.0
attrs==25.4.0
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
coloredlogs==15.0.1
cryptography==46.0.5
distro==1.9.0
exceptiongroup==1.3.1
flatbuffers==25.12.19
google-auth==2.48.0
google-genai==1.47.0
google-search-results==2.4.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
instaloader==4.15
jiter==0.13.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lazy_loader==0.4
llvmlite==0.43.0
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
onnxruntime==1.19.2
openai==2.24.0
opencv-python-headless==4.13.0.92
packaging==26.0
pillow==11.3.0
platformdirs==4.4.0
pooch==1.9.0
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyMatting==1.1.15
python-dotenv==1.2.1
referencing==0.36.2
rembg==2.0.61
requests==2.32.5
rpds-py==0.27.1
rsa==4.9.1
scikit-image==0.24.0
scipy==1.13.1
sniffio==1.3.1
soupsieve==2.8.3
sympy==1.14.0
tenacity==9.1.2
tifffile==2024.8.30
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==15.0.1
